src/snowflakecli/nextflow/commands.py
import typer
from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
from snowflake.cli.api.output.types import CommandResult, MessageResult
from snowflakecli.nextflow.workflow.commands import app as workflow_app

app = SnowTyperFactory(
    name="nextflow",
    help="Run Nextflow workflows in Snowpark Container Service",
)
app.add_typer(workflow_app)
