pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "snowflake-cli-nextflow-plugin"
requires-python = ">=3.10"
dependencies = [
    "snowflake-cli>=3.0.0"
]
version = "0.0.1"

[project.entry-points."snowflake.cli.plugin.command"]
nextflow = "snowflakecli.nextflow.plugin_spec"
